requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
Pillow==10.0.0
opencv-python==4.8.0.74
numpy==1.24.3
